+++ Changelog.md
@@ -25,6 +25,7 @@
 - Removing existing NewRelic refs from Monorepo
 - Update to use kube-rbac-proxy from ECR
 - Add customer-issuer to p14c-bootstrap configs
+- Add customer-issuer to PA Admin pingaccess-p1as-sso container
 
 _Changes:_
 
@@ -42,6 +43,7 @@ _Changes:_
 - [X] PDO-7616 Update PingDirectory resource limits
 - [X] PDO-7763 Cost Savings: NGINX  (public and private): Reduce CPU and mem requests
 - [X] PDO-8109 Reconcile Grafana differences with p1as-observability
+- [X] PDO-8157 Configure PA Admin API to add customer tenant for OAuth
 - [X] PDO-8372 Upgrade Kustomize to 5.5.0
 - [X] PDO-8383 EKS Managed Add-Ons: EBS CSI
 - [X] PDO-8727 Add a monitor and alert for kube-proxy or larger network issues within a cluster